# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: CDE Platform engineering
  description: |
    Common Data Environment API for construction and asset information.

    ## Tenant scoping

    Every request is scoped to exactly one tenant, and the tenant is taken **from the authenticated principal** — never from a path, query, header, or body parameter. There is no tenant selector on any endpoint, and supplying one has no effect. A resource belonging to another tenant is not visible: reading it by id returns `404`, not `403`, so the API does not confirm that an id exists somewhere else.

    ## Errors

    Every error response is an RFC 9457 Problem Details document with the media type `application/problem+json`. It carries a `traceId` that identifies the request in the server logs — quote it when reporting a problem. The `type` member is a stable relative URI identifying the kind of failure, and is safe to branch on when the status code alone is too coarse; `detail` is human-readable text whose wording may change, and is not.

    Validation failures return `422` with an `invalidFields` array naming each rejected field. A request body that cannot be parsed at all returns `400`, because no field can be named.

    ## Rate limits

...

    ## Tracing

    Every response carries `X-Trace-Id`. Send a W3C `traceparent` header and its trace id is adopted, so a client-side trace and the server-side logs share one identifier.
  license:
    name: Proprietary
  title: CDE Platform API
...
servers:
...
security:
- bearerAuth: []
...
paths:
  /api/ai/messages:
    post:
      description: |-
        Forwards the request to the configured model provider and returns its reply verbatim. The request and response bodies are the provider's own formats, which is why they are described here as opaque JSON rather than modelled: this endpoint does not interpret them.

        **The payload is forwarded as supplied.** The redaction layer that the data-handling rules require — allow-list field selection, personal-data pseudonymisation, and a hard refusal on classified content — is not implemented here yet, so a caller is responsible for what it sends. On a deployment where the provider may not be called at all, the feature is switched off rather than relied upon to filter.

        Requires authentication. Calls are chargeable and are attributed to the caller.
      operationId: sendAssistantMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantPayload"
              example: null
        description: "The provider's own request format, forwarded unchanged."
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPayload"
                example: null
          description: "The provider's reply, passed through unchanged."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPayload"
                example: null
          description: "The provider rejected the payload. Its own error format is\
            \ returned unchanged, so the members are the provider's, not this API's."
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No provider credential is configured on this deployment, or\
            \ the provider could not be reached."
      summary: Send a message to the assistant
      tags:
      - Assistance
  /api/ai/messages/stream:
    post:
      description: |-
        Same contract as the non-streaming route and, at present, the same behaviour: the reply is collected in full before it is returned, so a caller sees no incremental output. The separate path exists so streaming can be added behind it without a client change.

        The same caveat about unredacted payloads applies.

        Requires authentication. Calls are chargeable and are attributed to the caller.
      operationId: streamAssistantMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantPayload"
              example: null
        description: "The provider's own request format, forwarded unchanged."
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPayload"
                example: null
          description: "The provider's reply, passed through unchanged."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantPayload"
                example: null
          description: "The provider rejected the payload. Its own error format is\
            \ returned unchanged, so the members are the provider's, not this API's."
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No provider credential is configured on this deployment, or\
            \ the provider could not be reached."
      summary: "Send a message to the assistant, asking for a streamed reply"
      tags:
      - Assistance
...
components:
  schemas:
...
    AssistantPayload:
      type:
      - object
      additionalProperties: true
      description: "The model provider's own message format, passed through unchanged\
        \ in both directions. Its members are defined by the provider, not by this\
        \ API, and can change when the provider's version changes — so this API does\
        \ not enumerate them and does not validate them."
      example: null
...
    ProblemDetail:
      type:
      - object
      description: "RFC 9457 Problem Details. The single error shape returned by every\
        \ endpoint, for every status."
      example: null
      properties:
        detail:
          type:
          - string
          description: "What happened and what to do next, written for a person to\
            \ read. Never contains a stack trace, SQL, or an internal type name."
          example: 1 field is invalid.
        instance:
          type:
          - string
          format: uri-reference
          description: Path of the request that failed.
          example: /api/projects
        invalidFields:
          type:
          - array
          description: "Present on `422` validation failures. Names every field that\
            \ was rejected, so a client can attach each message to the control it\
            \ belongs to rather than to the form as a whole."
          example: null
          items:
            type:
            - object
            description: One field the request was rejected for.
            example: null
            properties:
              field:
                type:
                - string
                description: "Path of the rejected field, as it appears in the request\
                  \ body."
                example: name
              message:
                type:
                - string
                description: "What is wrong with the value, phrased for the person\
                  \ who typed it."
                example: must not be blank
            required:
            - field
            - message
        status:
          type:
          - integer
          format: int32
          description: "HTTP status code, repeated here so a logged body is self-contained."
          example: 422
          maximum: 599
          minimum: 400
        timestamp:
          type:
          - string
          format: date-time
          description: "When the failure was recorded, UTC."
          example: 2026-02-17T09:41:12.884Z
        title:
          type:
          - string
          description: Short summary of the problem type. Does not vary between occurrences.
          example: Validation failed
        traceId:
          type:
          - string
          description: "Identifies this request in the server logs, and is also returned\
            \ in the `X-Trace-Id` header. Quote it when reporting a problem."
          example: 4f8a1c2e9b7d6a5f3e2d1c0b9a8f7e6d
          pattern: "^[0-9a-f]{32}$"
        type:
          type:
          - string
          format: uri-reference
          description: "Stable identifier for this kind of problem, relative to the\
            \ API host. Safe to branch on — unlike `detail`, its value does not change\
            \ with wording."
          example: /problems/validation-failed
      required:
      - type
      - title
      - status
      - detail
      - traceId
...
  securitySchemes:
    apiKey:
      description: "Machine credential, sent as `X-API-Key`. Shown once at creation\
        \ and stored only as a SHA-256 digest, so a lost key is replaced rather than\
        \ recovered. Scoped, expiring, and rate-limited per key."
      in: header
      name: X-API-Key
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      description: "Signed JSON Web Token from `POST /api/auth/login`, sent as `Authorization:\
        \ Bearer <token>`. The token carries the subject and the tenant; the tenant\
        \ claim is authoritative and is not overridable by any request parameter."
      scheme: bearer
      type: http
    oauth2:
      description: Authorization code with PKCE for user-facing clients; client credentials
        for machine clients. The implicit flow is not offered.
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          refreshUrl: /oauth2/token
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
            project:write: "Create, update and delete projects"
            signature:read: Read and verify digital signatures
            signature:write: Apply a digital signature to a document version
          tokenUrl: /oauth2/token
        clientCredentials:
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
...